abstract flux specific resources
