+++ k8s-plugin/src/main/kotlin/com/amazon/spinnaker/keel/k8s/K8sConfig.kt
@@ -48,6 +48,7 @@ const val APPLICATION: String = "application"
 const val SECRET: String = "Secret"
 const val SECRET_API_V1: String = "v1"
 const val K8S_LIST: String = "List"
+const val K8S_LIST_API_V1: String = "List"
 const val FLUX_HELM_API_VERSION: String = "helm.toolkit.fluxcd.io/v2beta1"
 const val FLUX_HELM_KIND: String = "HelmRelease"
 const val FLUX_KUSTOMIZE_API_VERSION: String = "kustomize.toolkit.fluxcd.io/v1beta1"
