+++ k8s-plugin/src/main/kotlin/com/amazon/spinnaker/keel/k8s/model/BaseFluxArtifact.kt
@@ -1,0 +1,11 @@
+package com.amazon.spinnaker.keel.k8s.model
+
+import com.netflix.spinnaker.keel.api.artifacts.DeliveryArtifact
+
+abstract class BaseFluxArtifact(
+
+): DeliveryArtifact() {
+    abstract val kind: String
+    abstract val secretRef: String?
+    abstract val interval: String
+}
